Added crash protection on ID inputs. Simplified error messages.

+++ work-instructions-backend/controllers/items.js
@@ -1,5 +1,6 @@
 const Item = require("../models/Item");
 const { createCustomError } = require("../errors/custom-error");
+const checkId = require("../helpers/check-id");
 
 const getAllItems = async (req, res) => {
   const items = await Item.find({});
@@ -13,34 +14,38 @@ const createItem = async (req, res) => {
 
 const getItem = async (req, res, next) => {
   const { id: itemID } = req.params;
-  const item = await Item.findOne({ _id: itemID });
-  if (!item) {
-    return next(createCustomError(`No item with id : ${itemID}`, 404));
+
+  const validItemId = await checkId(itemID, Item);
+
+  if (!validItemId) {
+    return next(createCustomError(`No item with id: ${itemID}`, 404));
   }
+  const item = await Item.findOne({ _id: itemID });
 
   res.status(200).json({ item });
 };
 
 const deleteItem = async (req, res, next) => {
   const { id: itemID } = req.params;
-  const item = await Item.findOneAndDelete({ _id: itemID });
-  if (!item) {
-    return next(createCustomError(`No item with id : ${itemID}`, 404));
+  const validItemId = await checkId(itemID, Item);
+  if (!validItemId) {
+    return next(createCustomError(`No item with id: ${itemID}`, 404));
   }
+  const item = await Item.findOneAndDelete({ _id: itemID });
+
   res.status(200).json({ item });
 };
 
 const updateItem = async (req, res, next) => {
   const { id: itemID } = req.params;
-
+  const validItemId = await checkId(itemID, Item);
+  if (!validItemId) {
+    return next(createCustomError(`No item with id: ${itemID}`, 404));
+  }
   const item = await Item.findOneAndUpdate({ _id: itemID }, req.body, {
     new: true, // returns the new object
     runValidators: true,
   });
-
-  if (!item) {
-    return next(createCustomError(`No item with id : ${itemID}`, 404));
-  }
 
   res.status(200).json({ item });
 };
